backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.0
pydantic==2.5.3
pydantic-settings==2.1.0

# LangChain and LangGraph
langchain==0.1.0
langchain-openai==0.0.2
langchain-google-genai==0.0.6
langgraph==0.0.20
langchain-community==0.0.13

# Database
supabase==2.3.4
psycopg2-binary==2.9.9
asyncpg==0.29.0

# HTTP clients
httpx==0.26.0
aiohttp==3.9.1

# Search tools
duckduckgo-search==4.1.1
tavily-python==0.3.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
